ingest/requirements.txt
minio
opensearch-py
PyNaCl
geoip2
cryptography
pyyaml

